move Draco compression classes into the foxglove.remote_access namespace
DracoEncodeOptions and DracoMethod were registered on the top-level
foxglove module, so on lean wheels (built without the remote-access
feature) they were importable but unusable dead surface. Move them into
the foxglove.remote_access submodule alongside the rest of the gateway
API, so they are present exactly when remote access is.

The pyclasses move to remote_access.rs (module = "foxglove.remote_access")
and register via register_submodule; __init__.py re-exports
DracoEncodeOptions inside the guarded remote_access import for the
start_gateway annotation. Stubs, docs, and the option-class tests move to
their remote-access-specific homes.
