+++ python/foxglove-sdk/python/foxglove/_foxglove_py/__init__.pyi
@@ -25,6 +25,7 @@ class McapWritable(Protocol):
 from .mcap import MCAPWriteOptions, MCAPWriter
 from .remote_access import Capability as RemoteAccessCapability
 from .remote_access import (
+    DracoEncodeOptions,
     QosProfile,
     RemoteAccessConnectionStatus,
     RemoteAccessGateway,
@@ -161,33 +162,6 @@ class ChannelDescriptor:
     schema: "Schema" | None
 
 SinkChannelFilter = Callable[[ChannelDescriptor], bool]
-
-class DracoMethod(Enum):
-    """
-    Draco encoding method for point-cloud compression.
-
-    kd-tree is currently the only offered method; a ``Sequential`` value will be added
-    once an upstream encoder conformance bug is fixed.
-    """
-
-    KdTree = ...
-    """kd-tree encoding: reorders points, and float32 extra fields are quantized with the
-    same number of bits as positions. This is the default."""
-
-class DracoEncodeOptions:
-    """
-    Options for Draco point-cloud encoding.
-    """
-
-    method: DracoMethod
-    quantization_bits: int
-
-    def __init__(
-        self,
-        *,
-        method: DracoMethod = DracoMethod.KdTree,
-        quantization_bits: int = 12,
-    ) -> None: ...
 
 class ConnectionGraph:
     """
