+++ python/foxglove-sdk/src/remote_access.rs
@@ -12,10 +12,123 @@ use pyo3::types::PyBytes;
 use crate::PyContext;
 use crate::errors::PyFoxgloveError;
 use crate::logging::init_logging;
-use crate::remote_common::{
-    PyConnectionGraph, PyParameter, PyPointCloudCompression, PyService, PyStatusLevel,
-};
+use crate::remote_common::{PyConnectionGraph, PyParameter, PyService, PyStatusLevel};
 use crate::sink_channel_filter::{PyChannelDescriptor, PySinkChannelFilter};
+
+/// Draco encoding method for point-cloud compression.
+///
+/// kd-tree is currently the only offered method: sequential encoding is withheld because
+/// the encoder emits sequential bitstreams that the reference Draco decoder rejects
+/// whenever positions are quantized (upstream draco-core conformance bug). The enum and
+/// the ``method`` argument are kept so that re-offering sequential later is
+/// backwards-compatible; a ``Sequential`` value will be added once the upstream encoder
+/// is fixed.
+#[pyclass(
+    from_py_object,
+    name = "DracoMethod",
+    module = "foxglove.remote_access",
+    eq,
+    eq_int
+)]
+#[derive(Clone, PartialEq)]
+pub enum PyDracoMethod {
+    /// kd-tree encoding: reorders points, and float32 extra fields are quantized with the
+    /// same number of bits as positions. This is the default.
+    KdTree,
+}
+
+#[pymethods]
+impl PyDracoMethod {
+    #[getter]
+    fn name(&self) -> &'static str {
+        match self {
+            Self::KdTree => "KdTree",
+        }
+    }
+
+    #[getter]
+    fn value(&self) -> i32 {
+        match self {
+            Self::KdTree => 0,
+        }
+    }
+}
+
+/// Options for Draco point-cloud encoding.
+///
+/// :param method: The Draco encoding method. Currently kd-tree is the only choice; see
+///     :py:class:`DracoMethod`. Defaults to :py:attr:`DracoMethod.KdTree`.
+/// :type method: DracoMethod
+/// :param quantization_bits: Quantization bits for the position attribute; must be
+///     between 1 and 30 inclusive. Values outside that range are rejected when starting
+///     the gateway: values above 30 are rejected by the reference Draco decoder, and
+///     ``0`` (lossless) provides no size reduction over the raw point cloud — pass
+///     ``False`` as ``point_cloud_compression`` to disable compression instead.
+///     Defaults to 12.
+/// :type quantization_bits: int
+#[pyclass(
+    from_py_object,
+    name = "DracoEncodeOptions",
+    module = "foxglove.remote_access"
+)]
+#[derive(Clone)]
+pub struct PyDracoEncodeOptions {
+    #[pyo3(get, set)]
+    pub method: PyDracoMethod,
+    #[pyo3(get, set)]
+    pub quantization_bits: u8,
+}
+
+#[pymethods]
+impl PyDracoEncodeOptions {
+    #[new]
+    #[pyo3(signature = (*, method=PyDracoMethod::KdTree, quantization_bits=12))]
+    fn new(method: PyDracoMethod, quantization_bits: u8) -> Self {
+        Self {
+            method,
+            quantization_bits,
+        }
+    }
+}
+
+impl TryFrom<PyDracoEncodeOptions> for foxglove::draco::DracoEncodeOptions {
+    type Error = foxglove::FoxgloveError;
+
+    fn try_from(value: PyDracoEncodeOptions) -> Result<Self, Self::Error> {
+        // `method` has a single variant (kd-tree), which is also the only method the SDK
+        // derives for quantized encoding, so only the quantization setting is forwarded.
+        let PyDracoMethod::KdTree = value.method;
+        Self::with_quantization_bits(value.quantization_bits).map_err(|e| {
+            foxglove::FoxgloveError::ConfigurationError(format!("point_cloud_compression: {e}"))
+        })
+    }
+}
+
+/// The `point_cloud_compression` argument to `start_gateway`: either `DracoEncodeOptions`
+/// for custom settings, or a bool to enable (with default settings) or disable compression.
+#[derive(FromPyObject)]
+pub enum PyPointCloudCompression {
+    Draco(PyDracoEncodeOptions),
+    Enabled(bool),
+}
+
+impl PyPointCloudCompression {
+    /// Resolves the Python argument into the options a per-channel compression policy
+    /// applies, where `None` disables compression. Invalid Draco settings fail with a
+    /// configuration error, keeping misconfiguration a startup failure for Python
+    /// callers.
+    pub fn into_options(
+        self,
+    ) -> Result<Option<foxglove::draco::CompressPointCloudOptions>, foxglove::FoxgloveError> {
+        Ok(match self {
+            Self::Draco(options) => Some(foxglove::draco::CompressPointCloudOptions::Draco(
+                options.try_into()?,
+            )),
+            Self::Enabled(true) => Some(foxglove::draco::CompressPointCloudOptions::default()),
+            Self::Enabled(false) => None,
+        })
+    }
+}
 
 /// A client connected to a running remote access gateway.
 #[pyclass(name = "Client", module = "foxglove.remote_access")]
@@ -809,6 +922,8 @@ pub fn register_submodule(parent_module: &Bound<'_, PyModule>) -> PyResult<()> {
     module.add_class::<PyConnectionStatus>()?;
     module.add_class::<PyReliability>()?;
     module.add_class::<PyQosProfile>()?;
+    module.add_class::<PyDracoEncodeOptions>()?;
+    module.add_class::<PyDracoMethod>()?;
 
     let py = parent_module.py();
     py.import("sys")?
